+++ okapi-operation/src/axum_integration/handler_traits.rs
@@ -176,11 +176,14 @@ mod tests {
 
     use super::*;
     use crate::{
-        Components,
+        BuilderOptions, Components,
         axum_integration::{MethodRouter, Router},
     };
 
-    fn openapi_generator(_: &mut Components) -> Result<Operation, anyhow::Error> {
+    fn openapi_generator(
+        _: &mut Components,
+        _: &BuilderOptions,
+    ) -> Result<Operation, anyhow::Error> {
         unimplemented!()
     }